Refactoring, added comments.
