+++ package.json
@@ -14,6 +14,7 @@
     "cors": "^2.8.5",
     "express": "^4.18.2",
     "express-async-errors": "^3.1.1",
+    "express-rate-limit": "^6.7.0",
     "mysql2": "^3.1.2",
     "sync-mysql": "^3.0.1",
     "uuid": "^9.0.0"
