+++ records/category.record.ts
@@ -1,7 +1,7 @@
 import {ValidationError} from "../utils/error";
 import {pool} from "../utils/db";
 import {FieldPacket} from "mysql2";
-import {CategoryEntity, NewCategoryEntity} from "../types/category";
+import {CategoryEntity, NewCategoryEntity} from "../types";
 import {v4 as uuid} from "uuid";
 
 type CategoryRecordResult = [CategoryEntity[], FieldPacket[]];
@@ -13,7 +13,7 @@ export class CategoryRecord implements CategoryEntity {
     name: string;
     isDeletable: number;
 
-
+// validate new category object
     constructor(obj: NewCategoryEntity) {
         if (!obj.name || obj.name.length > 50) {
             throw new ValidationError('The name of category can not be empty or contains more than 50 chars.');
@@ -25,12 +25,13 @@ export class CategoryRecord implements CategoryEntity {
 
     }
 
-
+// listing all categories from db
     static async getAll(): Promise<CategoryEntity[]> {
         const [results] = await pool.execute('SELECT `category`.`id`, `category`.`name`, `category`.`is_deletable` AS isDeletable FROM `the_budget`.`category`') as CategoryRecordResult;
         return results.map(result => new CategoryRecord(result));
     }
 
+// deleting specific category from db
     static async deleteFromDb(id: string): Promise<CategoryRecordResult[]> {
         if (!id) {
             throw new Error(`Category with ${id} is not exist in database.`)
@@ -42,6 +43,7 @@ export class CategoryRecord implements CategoryEntity {
         return affectedRows as CategoryRecordResult[];
     }
 
+// adding new category to db
     async insertToDb(): Promise<string> {
         if (!this.id) {
             this.id = uuid();
@@ -53,6 +55,4 @@ export class CategoryRecord implements CategoryEntity {
         });
         return this.id;
     }
-
-
 }
