+++ records/shop.record.ts
@@ -7,14 +7,13 @@ import {v4 as uuid} from "uuid";
 type ShopRecordResult = [ShopEntity[], FieldPacket[]];
 
 
-
 export class ShopRecord implements ShopEntity {
 
     id?: string;
     name: string;
     isDeletable: number;
 
-
+// validation of new shop object
     constructor(obj: NewShopEntity) {
         if (!obj.name || obj.name.length > 50) {
             throw new ValidationError('The name can not be empty or contains more than 50 chars.');
@@ -26,13 +25,13 @@ export class ShopRecord implements ShopEntity {
 
     }
 
-
+// listing shops from db
     static async getAll(): Promise<ShopEntity[]> {
         const [results] = await pool.execute('SELECT `id`, `name`, `is_deletable` AS isDeletable FROM `the_budget`.`place`') as ShopRecordResult;
         return results.map(result => new ShopRecord(result));
     }
 
-
+// deleting specific shop from db
     static async deleteFromDb(id: string): Promise<ShopRecordResult[]> {
         if (!id) {
             throw new Error(`Shop with ${id} is not exist in database.`)
@@ -44,6 +43,7 @@ export class ShopRecord implements ShopEntity {
         return affectedRows as ShopRecordResult[];
     }
 
+// adding shop to db
     async insertToDb(): Promise<string> {
         if (!this.id) {
             this.id = uuid();
