+++ routers/budget.router.ts
@@ -6,39 +6,23 @@ import {BudgetRecord} from "../records/budget.record";
 export const budgetRouter = Router();
 
 budgetRouter
+
+    // getting budget and expense values from db
     .get('/showBudgetExpense', async (req, res) => {
         const budgetAndExpense: BudgetEntity[] = await BudgetRecord.getBudgetAndExpense();
         res.json(budgetAndExpense);
     })
+
+    // getting stats per category - [{categoryName: 'Food', expenseSum: 58}]
     .get('/getStatsPerCategory', async (req, res) => {
         const getStats: StatsFromBudget[] = await BudgetRecord.getStatsPerCategory();
         res.json(getStats);
     })
+
+    // updating changed value of budget
     .put('/', async (req, res) => {
         const newBudgetValue = await BudgetRecord.updateBudget(req.body.budget)
         res.json(newBudgetValue);
     });
-// .delete('/:id', async (req, res) => {
-//     const deleteShop = await ShopRecord.deleteFromDb(req.params.id);
-//     res.json(deleteShop);
-// })
-// .post('/', async (req, res) => {
-//     const addShop = new ShopRecord(req.body);
-//     console.log(addShop);
-//     await addShop.insertToDb();
-//     res.json(addShop);
-//
-// });
 
 
-// .get('/:id', async (req, res) => {
-//     const ad: AdEntity = await AdRecord.getOne(req.params.id);
-//     res.json(ad);
-// })
-//
-// .post('/', async (req, res) => {
-//     const ad = new AdRecord(req.body);
-//     await ad.insert();
-//     res.json(ad);
-//
-// });
