+++ utils/regexpFormatDate.ts
@@ -1,1 +1,2 @@
+//regexp for date format
 export const regexp = new RegExp(/([12]\d{3}-(0[1-9]|1[0-2])-(0[1-9]|[12]\d|3[01]))/);
